+++ src/RestaurantDetail.jsx
@@ -1,6 +1,8 @@
+import React from 'react';
+
 import MenuItems from './MenuItems';
 
-export default function RestaurantDetail({ restaurant }) {
+function RestaurantDetail({ restaurant }) {
   const { name, address, menuItems } = restaurant;
 
   return (
@@ -16,3 +18,5 @@ export default function RestaurantDetail({ restaurant }) {
     </div>
   );
 }
+
+export default React.memo(RestaurantDetail);
