+++ src/ReviewForm.jsx
@@ -1,6 +1,8 @@
+import React from 'react';
+
 import TextField from './TextField';
 
-export default function ReviewForm({ fields, onChange, onSubmit }) {
+function ReviewForm({ fields, onChange, onSubmit }) {
   const { score, description } = fields;
 
   return (
@@ -27,3 +29,5 @@ export default function ReviewForm({ fields, onChange, onSubmit }) {
     </>
   );
 }
+
+export default React.memo(ReviewForm);
